+++ pkg/render/helper/appfile.go
@@ -10,7 +10,7 @@ import (
 
 // RenderAppFile renders a single file using the current sequence of the provided app, or the overrideSequence (if provided)
 // it's here for now to avoid an import cycle between kotsadm/pkg/render and pkg/store
-func RenderAppFile(a *types.App, overrideSequence *int64, inputContent []byte, kotsKinds *kotsutil.KotsKinds, namespace string) ([]byte, error) {
+func RenderAppFile(a *types.App, overrideSequence *float64, inputContent []byte, kotsKinds *kotsutil.KotsKinds, namespace string) ([]byte, error) {
 	latestVersion, err := store.GetStore().GetLatestAppVersion(a.ID, true)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to get latest app version")
