+++ pkg/store/kotsstore/preflight_store.go
@@ -12,7 +12,7 @@ import (
 	troubleshootpreflight "github.com/replicatedhq/troubleshoot/pkg/preflight"
 )
 
-func (s *KOTSStore) SetPreflightProgress(appID string, sequence int64, progress string) error {
+func (s *KOTSStore) SetPreflightProgress(appID string, sequence float64, progress string) error {
 	db := persistence.MustGetDBSession()
 	query := `update app_downstream_version set preflight_progress = $1 where app_id = $2 and parent_sequence = $3`
 
@@ -24,7 +24,7 @@ func (s *KOTSStore) SetPreflightProgress(appID string, sequence int64, progress 
 	return nil
 }
 
-func (s *KOTSStore) GetPreflightProgress(appID string, sequence int64) (string, error) {
+func (s *KOTSStore) GetPreflightProgress(appID string, sequence float64) (string, error) {
 	db := persistence.MustGetDBSession()
 	query := `
 	SELECT preflight_progress
@@ -41,7 +41,7 @@ func (s *KOTSStore) GetPreflightProgress(appID string, sequence int64) (string, 
 	return progress.String, nil
 }
 
-func (s *KOTSStore) SetPreflightResults(appID string, sequence int64, results []byte) error {
+func (s *KOTSStore) SetPreflightResults(appID string, sequence float64, results []byte) error {
 	db := persistence.MustGetDBSession()
 	query := `update app_downstream_version set preflight_result = $1, preflight_result_created_at = $2,
 status = (case when status = 'deployed' then 'deployed' else 'pending' end),
@@ -56,7 +56,7 @@ where app_id = $3 and parent_sequence = $4`
 	return nil
 }
 
-func (s *KOTSStore) GetPreflightResults(appID string, sequence int64) (*preflighttypes.PreflightResult, error) {
+func (s *KOTSStore) GetPreflightResults(appID string, sequence float64) (*preflighttypes.PreflightResult, error) {
 	db := persistence.MustGetDBSession()
 	query := `
 	SELECT
@@ -83,7 +83,7 @@ func (s *KOTSStore) GetPreflightResults(appID string, sequence int64) (*prefligh
 	return r, nil
 }
 
-func (s *KOTSStore) ResetPreflightResults(appID string, sequence int64) error {
+func (s *KOTSStore) ResetPreflightResults(appID string, sequence float64) error {
 	db := persistence.MustGetDBSession()
 	query := `update app_downstream_version set preflight_result=null, preflight_result_created_at=null, preflight_skipped=false where app_id = $1 and parent_sequence = $2`
 	_, err := db.Exec(query, appID, sequence)
@@ -93,7 +93,7 @@ func (s *KOTSStore) ResetPreflightResults(appID string, sequence int64) error {
 	return nil
 }
 
-func (s *KOTSStore) SetIgnorePreflightPermissionErrors(appID string, sequence int64) error {
+func (s *KOTSStore) SetIgnorePreflightPermissionErrors(appID string, sequence float64) error {
 	db := persistence.MustGetDBSession()
 	query := `UPDATE app_downstream_version
 	SET status = 'pending_preflight', preflight_ignore_permissions = true, preflight_result = null, preflight_skipped = false
